codex: address streamable http retry review feedback (#25147)
